Prefill Previous Exercise Values
When adding an exercise to a workout session,
and adding the first set, it will prefill the first sets values with the previous workout they did that exercise. After you fill in the value of the first set, the next sets added, will prefill what you entered on set one.